Fix issue with gcc ignoring -ldl flag
With automake, having -ldl in LDFLAGS causes the flags to be placed
before the objects during the link step. However, with gcc, the -ldl is
ignored because it is not needed until the objects are parsed. This is
not an issue with clang.

Moving the -ldl to LDADD works around gcc behavior to add the flag at
the end of the command line.